+++ lib/rubocop.rb
@@ -83,6 +83,7 @@ require 'rubocop/cop/end_alignment'
 require 'rubocop/cop/void'
 require 'rubocop/cop/documentation'
 require 'rubocop/cop/unreachable_code'
+require 'rubocop/cop/method_invocation_parentheses'
 
 require 'rubocop/cop/rails/validation'